fix Menu Managemer, update status menu

+++ Repositories/Interfaces/IMenuItemRepository.cs
@@ -6,8 +6,8 @@ namespace Repositories.Interfaces
     public interface IMenuItemRepository
     {
         Task<IReadOnlyList<MenuItem>> GetAllMenuItemsAsync(CancellationToken ct = default);
-        Task<MenuItem> GetMenuItemByIdAsync(int id, CancellationToken ct = default);
-        Task<bool> DeleteMenuItemAsync(int dishId, CancellationToken ct = default);
+        Task<MenuItem> GetMenuItemByIdAsync(long id, CancellationToken ct = default);
+        Task<bool> DeleteMenuItemAsync(long dishId, CancellationToken ct = default);
         Task<bool> AddMenuItemAsync(MenuItem item, CancellationToken ct = default);
         Task<bool> UpdateMenuItemAsync(MenuItem menuItem, CancellationToken ct);
     }
